Unable to change Court Grants [#79 state:fixed responsible:none milestone:v2.1.2]
Fixed in changeset11. The LRS migration did not bring the court grant Folio reference
correctly into SOLA. Have updated so that all Court Grants now have the Lot/Parcel Number
of CTGT and the Plan number is the grant number. E.g. CTGT/593. This will allow them to
be added as Properties to the application.